# apps/api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/resources/{id}": {
      "get": {
        "operationId": "ResourcesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resources"
        ]
      },
      "put": {
        "operationId": "ResourcesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resources"
        ]
      },
      "delete": {
        "operationId": "ResourcesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "tags": [
          "Resources"
        ]
      }
    },
    "/resource-types": {
      "get": {
        "operationId": "ResourceTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceTypes"
        ]
      },
      "post": {
        "operationId": "ResourceTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ResourceTypes"
        ]
      }
    },
    "/resource-types/{id}": {
      "get": {
        "operationId": "ResourceTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "ResourceTypes"
        ]
      },
      "patch": {
        "operationId": "ResourceTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceTypes"
        ]
      },
      "delete": {
        "operationId": "ResourceTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ResourceTypes"
        ]
      }
    },
...
  },
  "info": {
    "title": "File storage API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category_id": {
            "type": "number"
          },
          "publication_date": {
            "format": "date-time",
            "type": "string"
          },
          "authors": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keywords": {
            "default": [],
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "file_name": {
            "type": "string"
          },
          "file_format": {
            "type": "object"
          },
          "file_size": {
            "type": "string"
          },
          "preview_image_name": {
            "type": "string",
            "format": "uri"
          },
          "citation": {
            "type": "string"
          },
          "additional_info": {
            "type": "object",
            "nullable": true
          },
          "resource_type_id": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "category_id",
          "publication_date",
          "authors",
          "keywords",
          "file_name",
          "file_format",
          "file_size",
          "resource_type_id"
        ]
      },
      "CreateResourceTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "UpdateResourceTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    }
  }
}